SNTWREC-327-Refinement-Widget
Introducing a new widget that enables the users to choose the type of how they refine the results.
The widget is a radio box inside an accordion and has three options ['Similar', 'Diverse', and 'Recent] and each type has two buttons to change the results either more or less.
The buttons are disabled when we reach a certain threshold limit from the back-end and an alert is then shown.
going the other direction will hide the alert message and reactive the deactivated button.

The first default call is used with the similarity type and it uses the default weights in the PA-Service "which to be changed in later tickets" and then the following calls will include the previously received weights and ids of the recommendations so that we can calculate new recommendations in the back-end.

This ticket also includes using three endpoints where each is linked to a specific type of refinement and if the widget was never used in the config then Reco will behave as before calling the legacy endpoint for WDR and the similarity endpoint only for BR